+++ imod/mf6/rch.py
@@ -6,12 +6,16 @@ import numpy as np
 import xarray as xr
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
-from imod.common.utilities.regrid import _regrid_package_data
+from imod.common.utilities.dataclass_type import DataclassType
 from imod.logging import init_log_decorator
+from imod.mf6.aggregate.aggregate_schemes import RechargeAggregationMethod
 from imod.mf6.boundary_condition import BoundaryCondition
-from imod.mf6.dis import StructuredDiscretization
+from imod.mf6.dis import StructuredDiscretization, VerticesDiscretization
 from imod.mf6.regrid.regrid_schemes import RechargeRegridMethod
-from imod.mf6.utilities.imod5_converter import convert_unit_rch_rate
+from imod.mf6.utilities.imod5_converter import (
+    convert_unit_rch_rate,
+    regrid_imod5_pkg_data,
+)
 from imod.mf6.utilities.package import set_repeat_stress_if_available
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.msw.utilities.imod5_converter import (
@@ -133,6 +137,7 @@ class Recharge(BoundaryCondition, IRegridPackage):
     _template = BoundaryCondition._initialize_template(_pkg_id)
     _auxiliary_data = {"concentration": "species"}
     _regrid_method = RechargeRegridMethod()
+    _aggregate_method: DataclassType = RechargeAggregationMethod()
 
     @init_log_decorator()
     def __init__(
@@ -173,7 +178,8 @@ class Recharge(BoundaryCondition, IRegridPackage):
     def allocate_planar_data(
         cls,
         planar_data: dict[str, GridDataArray],
-        dis: StructuredDiscretization,
+        dis: StructuredDiscretization | VerticesDiscretization,
+        allocation_option: ALLOCATION_OPTION,
     ) -> dict[str, GridDataArray]:
         """
         Allocate and distribute planar data for given discretization and npf
@@ -186,8 +192,8 @@ class Recharge(BoundaryCondition, IRegridPackage):
             Dictionary with planar grid data.
         dis: imod.mf6.StructuredDiscretization
             Model discretization package.
-        npf: imod.mf6.NodePropertyFlow
-            Node property flow package.
+        allocation_option: ALLOCATION_OPTION
+            The allocation option to use for the reallocation.
 
         Returns
         -------
@@ -199,7 +205,7 @@ class Recharge(BoundaryCondition, IRegridPackage):
             planar_data["rate"] = planar_data["rate"].isel(layer=0, drop=True)
         # create an array indicating in which cells rch is active
         is_rch_cell = allocate_rch_cells(
-            ALLOCATION_OPTION.at_first_active,
+            allocation_option,
             idomain > 0,
             planar_data["rate"],
         )
@@ -255,21 +261,18 @@ class Recharge(BoundaryCondition, IRegridPackage):
         Modflow 6 rch package.
 
         """
-        new_idomain = target_dis.dataset["idomain"]
         data = {
             "rate": convert_unit_rch_rate(imod5_data["rch"]["rate"]),
         }
-        # first regrid the inputs to the target grid.
-        if regridder_types is None:
-            regridder_settings = Recharge.get_regrid_methods()
-
-        regridded_package_data = _regrid_package_data(
-            data, new_idomain, regridder_settings, regrid_cache, {}
+        regridded_package_data = regrid_imod5_pkg_data(
+            cls, data, target_dis, regridder_types, regrid_cache
         )
-
         # if rate has only layer 0, then it is planar.
         if is_planar_grid(regridded_package_data["rate"]):
-            layered_data = cls.allocate_planar_data(regridded_package_data, target_dis)
+            allocation_option = ALLOCATION_OPTION.at_first_active
+            layered_data = cls.allocate_planar_data(
+                regridded_package_data, target_dis, allocation_option
+            )
             regridded_package_data.update(layered_data)
         rch = cls(**regridded_package_data, validate=True, fixed_cell=False)
         repeat = period_data.get("rch")