+++ imod/mf6/utilities/imod5_converter.py
@@ -4,8 +4,8 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
+from imod.common.utilities.dataclass_type import DataclassType
 from imod.common.utilities.regrid import _regrid_package_data
-from imod.common.utilities.regrid_method_type import RegridMethodType
 from imod.mf6.package import Package
 from imod.typing import GridDataArray, GridDataDict, Imod5DataDict
 from imod.typing.grid import full_like
@@ -66,7 +66,10 @@ def _well_from_imod5_cap_point_data(cap_data: GridDataDict) -> dict[str, np.ndar
 def _well_from_imod5_cap_grid_data(cap_data: GridDataDict) -> dict[str, np.ndarray]:
     artificial_rch_type = cap_data["artificial_recharge"]
     layer = cap_data["artificial_recharge_layer"].astype(int)
-
+    # Workaround to be able to later specify groundwater abstraction rates as
+    # well as surface water abstraction rates in a single row, which is a
+    # requirement of MetaSWAP. If artificial_rch_type == 1 (groundwater
+    # abstraction), the surface water abstraction wells are set to 0.0 later.
     from_groundwater = (artificial_rch_type != 0).to_numpy()
     coords = artificial_rch_type.coords
     x_grid, y_grid = np.meshgrid(coords["x"].to_numpy(), coords["y"].to_numpy())
@@ -119,7 +122,7 @@ def regrid_imod5_pkg_data(
     cls: type[Package],
     imod5_pkg_data: GridDataDict,
     target_dis: Package,
-    regridder_types: Optional[RegridMethodType],
+    regridder_types: Optional[DataclassType],
     regrid_cache: RegridderWeightsCache,
 ) -> GridDataDict:
     """